refactor: move time preference policy into application
