+++ src/lifeos_cli/db/services/events.py
@@ -9,6 +9,7 @@ from sqlalchemy import or_, select
 from sqlalchemy.ext.asyncio import AsyncSession
 from sqlalchemy.orm import selectinload
 
+from lifeos_cli.application.time_preferences import get_utc_window_for_local_date
 from lifeos_cli.db.models.event import Event
 from lifeos_cli.db.models.person_association import person_associations
 from lifeos_cli.db.models.tag_association import tag_associations
@@ -29,7 +30,6 @@ from lifeos_cli.db.services.event_support import (
     validate_event_time_range,
     validate_event_title,
 )
-from lifeos_cli.time_preferences import get_utc_window_for_local_date
 
 
 async def _attach_event_links(session: AsyncSession, event: Event) -> Event:
